[BEAM-6188] Move SyntheticBoundedSource to upper level
